third_party/WebKit/LayoutTests/fast/dom/geometry-interfaces-dom-matrix-expected.txt

# DOMMatrix()
PASS matrix.a is matrix.m11
PASS matrix.b is matrix.m12
PASS matrix.c is matrix.m21
PASS matrix.d is matrix.m22
PASS matrix.e is matrix.m41
PASS matrix.f is matrix.m42
PASS matrix.m11 is 1
PASS matrix.m12 is 0
PASS matrix.m13 is 0
PASS matrix.m14 is 0
PASS matrix.m21 is 0
PASS matrix.m22 is 1
PASS matrix.m23 is 0
PASS matrix.m24 is 0
PASS matrix.m31 is 0
PASS matrix.m32 is 0
PASS matrix.m33 is 1
PASS matrix.m34 is 0
PASS matrix.m41 is 0
PASS matrix.m42 is 0
PASS matrix.m43 is 0
PASS matrix.m44 is 1
PASS matrix.is2D is true
PASS matrix.isIdentity is true

# DOMMatrix set attributes
PASS matrix.a is matrix.m11
PASS matrix.b is matrix.m12
PASS matrix.c is matrix.m21
PASS matrix.d is matrix.m22
PASS matrix.e is matrix.m41
PASS matrix.f is matrix.m42
PASS matrix.m11 is 10
PASS matrix.m12 is 20
PASS matrix.m13 is 0
PASS matrix.m14 is 0
PASS matrix.m21 is 0
PASS matrix.m22 is 1
PASS matrix.m23 is 0
PASS matrix.m24 is 2
PASS matrix.m31 is 0
PASS matrix.m32 is 0
PASS matrix.m33 is 3
PASS matrix.m34 is 0
PASS matrix.m41 is 0
PASS matrix.m42 is 3
PASS matrix.m43 is 0
PASS matrix.m44 is 9
PASS matrix.is2D is false
PASS matrix.isIdentity is false

# DOMMatrix(other)
PASS matrix.a is matrix.m11
PASS matrix.b is matrix.m12
PASS matrix.c is matrix.m21
PASS matrix.d is matrix.m22
PASS matrix.e is matrix.m41
PASS matrix.f is matrix.m42
PASS matrix.m11 is 10
PASS matrix.m12 is 20
PASS matrix.m13 is 0
PASS matrix.m14 is 0
PASS matrix.m21 is 0
PASS matrix.m22 is 1
PASS matrix.m23 is 0
PASS matrix.m24 is 2
PASS matrix.m31 is 0
PASS matrix.m32 is 0
PASS matrix.m33 is 3
PASS matrix.m34 is 0
PASS matrix.m41 is 0
PASS matrix.m42 is 3
PASS matrix.m43 is 0
PASS matrix.m44 is 9
PASS matrix.is2D is false
PASS matrix.isIdentity is false

# DOMMatrix.is2D can never be set to 'true' when it was set to 'false' before calling setMatrixValue().
PASS matrix.is2D is true
PASS matrix.isIdentity is true
PASS matrix.is2D is false
PASS matrix.isIdentity is false
PASS matrix.is2D is false
PASS matrix.isIdentity is true

PASS successfullyParsed is true

TEST COMPLETE
